reverse engineered to get ER diagram

+++ rds/structure/shikshithadb_sub_album.sql
@@ -23,7 +23,7 @@ DROP TABLE IF EXISTS `sub_album`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `sub_album` (
-  `Id` int(11) NOT NULL AUTO_INCREMENT,
+  `Id` bigint(20) NOT NULL AUTO_INCREMENT,
   `Name` varchar(50) NOT NULL,
   `AlbumId` bigint(20) NOT NULL,
   `CreatedBy` bigint(20) NOT NULL,
@@ -43,4 +43,4 @@ CREATE TABLE `sub_album` (
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2017-10-26 21:00:19
+-- Dump completed on 2017-10-29  8:32:10
